refactor: Move fetch method into service
